requirements.txt
astroid==2.2.5
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
mysql-connector==2.2.9
pylint==2.3.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
Unidecode==1.1.1
urllib3==1.25.3
wrapt==1.11.1
